+++ android/quest/src/main/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientDetailViewModel.kt
@@ -100,7 +100,7 @@ class QuestPatientDetailViewModel(
   }
 
   companion object {
-    const val PROFILE_CONFIG = "profile_config.json"
+    const val PROFILE_CONFIG = "configurations/form/profile_config.json"
 
     fun get(
       owner: ViewModelStoreOwner,