+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ReferenceAttachmentResolverTest.kt
@@ -17,10 +17,9 @@
 package org.smartregister.fhircore.quest
 
 import android.graphics.Bitmap
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.core.app.ApplicationProvider
-import com.google.android.fhir.FhirEngine
 import io.mockk.coEvery
-import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.mockkObject
@@ -30,6 +29,7 @@ import io.mockk.verify
 import java.io.ByteArrayInputStream
 import java.nio.charset.Charset
 import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.test.runBlockingTest
 import okhttp3.MediaType
 import okhttp3.ResponseBody
 import okio.BufferedSource
@@ -37,25 +37,27 @@ import org.hl7.fhir.r4.model.Binary
 import org.junit.After
 import org.junit.Assert
 import org.junit.Before
+import org.junit.Rule
 import org.junit.Test
-import org.robolectric.annotation.Config
-import org.robolectric.util.ReflectionHelpers
 import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceService
-import org.smartregister.fhirecore.quest.robolectric.RobolectricTest
-import org.smartregister.fhirecore.quest.shadow.QuestApplicationShadow
+import org.smartregister.fhircore.quest.coroutine.CoroutineTestRule
+import org.smartregister.fhircore.quest.robolectric.RobolectricTest
 import retrofit2.Call
 import retrofit2.Response
 
-@Config(shadows = [QuestApplicationShadow::class])
 class ReferenceAttachmentResolverTest : RobolectricTest() {
+  @get:Rule val instantTaskExecutorRule = InstantTaskExecutorRule()
 
-  private lateinit var referenceAttachmentResolverTest: ReferenceAttachmentResolver
+  @get:Rule val coroutinesTestRule = CoroutineTestRule()
+
+  private lateinit var referenceAttachmentResolver: ReferenceAttachmentResolver
+
   private lateinit var fhirService: FhirResourceService
 
   @Before
   fun setUp() {
-    referenceAttachmentResolverTest =
-      ReferenceAttachmentResolver(ApplicationProvider.getApplicationContext<QuestApplication>())
+    referenceAttachmentResolver =
+      ReferenceAttachmentResolver(ApplicationProvider.getApplicationContext())
 
     fhirService = mockk()
     mockkObject(FhirResourceService.Companion)
@@ -69,28 +71,19 @@ class ReferenceAttachmentResolverTest : RobolectricTest() {
 
   @Test
   fun testResolveBinaryResourceShouldCallFhirEngineLoadAndReturnBinary() {
-    val fhirEngine = mockk<FhirEngine>()
-    val expectedBinaryResource = Binary()
-
-    ReflectionHelpers.setField(
-      QuestApplication.getContext(),
-      "fhirEngine\$delegate",
-      lazy { fhirEngine }
-    )
-
-    coEvery { fhirEngine.load(Binary::class.java, "sample-binary-image") } returns
-      expectedBinaryResource
-
-    val actualBinaryResource: Binary?
-    runBlocking {
-      actualBinaryResource =
-        referenceAttachmentResolverTest.resolveBinaryResource(
+    coroutinesTestRule.runBlockingTest {
+      val expectedBinaryResource = Binary().apply { id = "binaryId" }
+      coEvery {
+        ApplicationProvider.getApplicationContext<QuestApplication>()
+          .fhirEngine
+          .load(Binary::class.java, "sample-binary-image")
+      } returns expectedBinaryResource
+      val actualBinaryResource =
+        referenceAttachmentResolver.resolveBinaryResource(
           "https://fhir-server.org/Binary/sample-binary-image"
         )
+      Assert.assertEquals(expectedBinaryResource, actualBinaryResource)
     }
-
-    coVerify { fhirEngine.load(Binary::class.java, "sample-binary-image") }
-    Assert.assertEquals(expectedBinaryResource, actualBinaryResource)
   }
 
   @Test
@@ -103,7 +96,7 @@ class ReferenceAttachmentResolverTest : RobolectricTest() {
     every { okHttpCall.execute() } returns mockResponse
     every { fhirService.fetchImage(any()) } returns okHttpCall
 
-    runBlocking { referenceAttachmentResolverTest.resolveImageUrl(imageUrl) }
+    runBlocking { referenceAttachmentResolver.resolveImageUrl(imageUrl) }
 
     verify { fhirService.fetchImage(imageUrl) }
   }
@@ -119,7 +112,7 @@ class ReferenceAttachmentResolverTest : RobolectricTest() {
     every { fhirService.fetchImage(imageUrl) } returns okHttpCall
 
     val bitmap: Bitmap?
-    runBlocking { bitmap = referenceAttachmentResolverTest.resolveImageUrl(imageUrl) }
+    runBlocking { bitmap = referenceAttachmentResolver.resolveImageUrl(imageUrl) }
 
     Assert.assertNull(bitmap)
   }
@@ -141,7 +134,7 @@ class ReferenceAttachmentResolverTest : RobolectricTest() {
     every { fhirService.fetchImage(imageUrl) } returns okHttpCall
 
     val bitmap: Bitmap?
-    runBlocking { bitmap = referenceAttachmentResolverTest.resolveImageUrl(imageUrl) }
+    runBlocking { bitmap = referenceAttachmentResolver.resolveImageUrl(imageUrl) }
 
     Assert.assertNotNull(bitmap)
   }
@@ -150,7 +143,7 @@ class ReferenceAttachmentResolverTest : RobolectricTest() {
   fun testGetContext() {
     Assert.assertEquals(
       ApplicationProvider.getApplicationContext(),
-      referenceAttachmentResolverTest.context
+      referenceAttachmentResolver.application
     )
   }